feat: tag-driven release pipeline
Pushing `vX.Y.Z` now verifies the tag against the workspace version, re-runs
the branch gate on the tagged commit across Linux and macOS at stable and the
MSRV, generates the changelog from the conventional-commit history, builds a
CLI binary per platform, publishes a GitHub release with checksums, and
fast-forwards main. Every step converges on re-runs, so a partial release is
finished by pushing the same tag again.

There is no crates.io publish, because a git tag is the distribution artifact.
The `git-dep` job is what publishing would be for a registry project: it builds
a scratch crate outside the workspace against `git = <repo>, tag = <tag>`, so
the instruction in README.md and API-POLICY.md is proven against the tag being
cut rather than assumed. Validated locally against a real tagged repository:
two packages resolve, no transitive dependencies.

The matrix sets RUSTUP_TOOLCHAIN explicitly. rust-toolchain.toml pins 1.89.0
and outranks `rustup default`, so without it every "stable" leg would silently
run the pinned toolchain and the matrix would test one version twice. A test
now pins that, and pins the gate command list across ci.yml, release.yml and
docs/RELEASE.md so the three copies cannot drift apart silently.

Making the matrix green required six real fixes surfaced by the stable
toolchain and by denying warnings: redundant u128 casts in the rescale path,
two is_some/unwrap pairs, two manual is_multiple_of implementations, and unused
imports in two tests.

The chaos harness now reports a classification histogram and covers the
torn-tail path by construction — truncate mid-record, classify, repair, refold
— instead of asserting that 1,000 random kills happened to produce one. They
produce about five on 1.89.0 and none on stable, because SIGKILL cannot
interrupt a write syscall to a regular file; the assertion was a coin flip that
proved nothing on the runs where it came up zero. The random kills and their
per-iteration recovery checks are unchanged.

+++ crates/fsm-core/src/decimal/mod.rs
@@ -295,8 +295,8 @@ impl Dec {
         let r = mag % div;
         let mut q = q;
         if r != 0 {
-            let twice = (r as u128).saturating_mul(2);
-            let ord = twice.cmp(&(div as u128));
+            let twice = r.saturating_mul(2);
+            let ord = twice.cmp(&div);
             if bump(mode, negative, ord, q.is_multiple_of(2)) {
                 q = q.checked_add(1).ok_or(DecError::Overflow)?;
             }
